requirements.txt
viennarna
biopython
numpy
pandas
tqdm
matplotlib
seaborn